Displaying all the Rocket data is finihsed

+++ app/src/main/java/com/robelseyoum3/safetonet/model/Engines.kt
@@ -19,7 +19,7 @@ data class Engines(
     @SerializedName("thrust_sea_level")
     val thrustSeaLevel: ThrustSeaLevel,
     @SerializedName("thrust_to_weight")
-    val thrustToWeight: Int,
+    val thrustToWeight: Double,
     @SerializedName("thrust_vacuum")
     val thrustVacuum: ThrustVacuum,
     @SerializedName("type")
