+++ app/src/main/java/com/robelseyoum3/safetonet/model/Height.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Height(
     @SerializedName("feet")
-    val feet: Int,
+    val feet: Double,
     @SerializedName("meters")
     val meters: Double
 )
